Migrate `/dashboard` to react
I like react. I think it can be an easy way to improve the frontend to a
presentable level.

With react also a build step is introduced which uses `babel` and
`webpack`. The nodejs buildpack is added to heroku to be able to execute
these.

Fixed versions in the `requirements.txt` to better track dependency
updates and have explicit dependency versions in the deploy step.

Add endpoint to fetch repositories for the dashboard.