Smaller fixes
Remove codacy push, this needs an token, which shouldn't be exposed to
external Pull Request. Remove for now to be able to include CI. Think
about a better solution later.

Fix the message formatting on reviewed pull requests. Add some logging
to be able to debug the vote calculation error better.
Addresses #162

Add `sort-package-json` in the CI checks. Due to Hacktoberfest there was
one Pull Request which just changes the package.json ordering. This can
make sense, but should be agreed on and enforced. So thx for idea.