Show PullRequests on the dashboard
Before on the dashboard only the repositories were shown with a checkbox
to enable World Driven and a list of Pull Requests.

Now a Pull Request can be selected and the details of the Pull Request
are shown.

This is in many regards (styling, code structure, ...) not close to
optimal, but it improves the current dashboard.

Next steps:

- Look into ways to include component css into the component
- Change the 'routing', put pull requests under `/dashboard`
(`/dashboard/:org/:repo/pull/:pull_number`) also change when selecting
PR. Get rid of the custom Pull Request logic.
- Better naming of components