Migrate database backend to PostgreSQL
The MongoDB addon used on heroku will shutdown soonish.
Instead of MongoDB PostgreSQL is now used.
Fixes #192

Flask-SQLAlchemy as ORM.
Flask-Migrate handling the migration

On app startup the existing data is automatically migrated.

The github access keys are now stored encrypted, the encryption key is
provided as `ENCRYPTION_KEY` as environment variable.
Fixes #197

Fixes #49 - The session are and will be stored in the database