Fix production issues
With `--chdir` the working directory is changed. This collides with
finding the migrations for sqlalchemy.
Changed to `--pythonpath` which only changes the python environment but
not the working directory.
Use correct environment variable.
Change to mysql/mariadb as this is used in production.
Run the regular job in the app context.

Yeah, I messed up properly with the production release, was a hit in a
hurry because of the MongoDB shutdown.