Fix production issues (#219)
* Fix production issues

With `--chdir` the working directory is changed. This collides with
finding the migrations for sqlalchemy.
Changed to `--pythonpath` which only changes the python environment but
not the working directory.
Use correct environment variable.
Change to mysql/mariadb as this is used in production.
Run the regular job in the app context.

Yeah, I messed up properly with the production release, was a hit in a
hurry because of the MongoDB shutdown.

* Change circle ci to mariadb container

* Remove mongodb

* Use 127.0.0.1

* Fix URL for worker process

Co-authored-by: Tobias Wilken <tw@tw-t450s>